refact: split multimodule parser into components
