+++ src/parser/mod.rs
@@ -5,4 +5,5 @@ mod parse_error;
 mod single_module;
 
 pub use facade::Parser;
+pub use fs::{FileSystem, FileSystemError};
 pub use parse_error::{ParseError, ParseErrorType};
