+++ src/parser/multi_module/implementation/mod.rs
@@ -1,0 +1,5 @@
+mod parser;
+mod path_collector_visitor;
+
+pub use parser::*;
+pub use path_collector_visitor::*;
