+++ src/parser/multi_module/mod.rs
@@ -1,7 +1,7 @@
-mod implemenation;
+mod implementation;
 mod interface;
 #[cfg(test)]
 mod tests;
 
-pub use implemenation::MultiModuleParserImpl;
+pub use implementation::MultiModuleParserImpl;
 pub use interface::MultiModuleParser;
